+++ pkg/compactor/compactor_metrics_test.go
@@ -130,6 +130,11 @@ func TestSyncerMetrics(t *testing.T) {
 			cortex_compactor_compaction_error_total{type="unauthorized",user="aaa"} 477730
 			cortex_compactor_compaction_error_total{type="unauthorized",user="bbb"} 488840
 			cortex_compactor_compaction_error_total{type="unauthorized",user="ccc"} 499950
+			# HELP cortex_compactor_group_partition_count Number of partitions for each compaction group.
+			# TYPE cortex_compactor_group_partition_count gauge
+			cortex_compactor_group_partition_count{user="aaa"} 511060
+			cortex_compactor_group_partition_count{user="bbb"} 522170
+			cortex_compactor_group_partition_count{user="ccc"} 533280
 	`))
 	require.NoError(t, err)
 
@@ -183,4 +188,7 @@ func generateTestData(cm *compactorMetrics, base float64) {
 	cm.compactionErrorsCount.WithLabelValues("aaa", unauthorizedError).Add(43 * base)
 	cm.compactionErrorsCount.WithLabelValues("bbb", unauthorizedError).Add(44 * base)
 	cm.compactionErrorsCount.WithLabelValues("ccc", unauthorizedError).Add(45 * base)
+	cm.partitionCount.WithLabelValues("aaa").Add(46 * base)
+	cm.partitionCount.WithLabelValues("bbb").Add(47 * base)
+	cm.partitionCount.WithLabelValues("ccc").Add(48 * base)
 }